+++ CHANGELOG.md
@@ -9,6 +9,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+## v6.63.0
+
 ### Added
 
 - Add a file based query cache to leverage OPcache for improved performance https://github.com/nuwave/lighthouse/pull/2713